feat: refactor footer and helper component
